+++ internal/service/s3/errors.go
@@ -4,12 +4,12 @@ package s3
 // https://docs.aws.amazon.com/sdk-for-go/api/service/s3/#pkg-constants
 
 const (
-	ErrCodeInvalidBucketState                   = "InvalidBucketState"
-	ErrCodeMethodNotAllowed                     = "MethodNotAllowed"
-	ErrCodeNoSuchConfiguration                  = "NoSuchConfiguration"
-	ErrCodeNoSuchCORSConfiguration              = "NoSuchCORSConfiguration"
-	ErrCodeNoSuchPublicAccessBlockConfiguration = "NoSuchPublicAccessBlockConfiguration"
-	ErrCodeNoSuchWebsiteConfiguration           = "NoSuchWebsiteConfiguration"
-	ErrCodeObjectLockConfigurationNotFound      = "ObjectLockConfigurationNotFoundError"
-	ErrCodeOperationAborted                     = "OperationAborted"
+	ErrCodeInvalidBucketState                        = "InvalidBucketState"
+	ErrCodeNoSuchConfiguration                       = "NoSuchConfiguration"
+	ErrCodeNoSuchCORSConfiguration                   = "NoSuchCORSConfiguration"
+	ErrCodeNoSuchPublicAccessBlockConfiguration      = "NoSuchPublicAccessBlockConfiguration"
+	ErrCodeNoSuchWebsiteConfiguration                = "NoSuchWebsiteConfiguration"
+	ErrCodeObjectLockConfigurationNotFound           = "ObjectLockConfigurationNotFoundError"
+	ErrCodeOperationAborted                          = "OperationAborted"
+	ErrCodeServerSideEncryptionConfigurationNotFound = "ServerSideEncryptionConfigurationNotFoundError"
 )